+++ Assignment 1/model/ProfessionalRunner.cs
@@ -4,8 +4,8 @@ public class ProfessionalRunner : Runner
 {
     public int WorldRanking { get; set; }
 
-    public ProfessionalRunner(string fullName, string address, int phoneNumber, string emailAddress, RankType rankType,
-        Status status, int number, int worldRanking) : base(fullName, address, phoneNumber, emailAddress, rankType,
+    public ProfessionalRunner(string fullName, string address, long phoneNumber, string emailAddress, RankType rankType,
+        Status status, long number, int worldRanking) : base(fullName, address, phoneNumber, emailAddress, rankType,
         status, number)
     {
         WorldRanking = worldRanking;
