+++ Assignment 1/model/Runner.cs
@@ -4,11 +4,11 @@ public class Runner : Participant
 {
     public RankType RankType { get; set; }
     public Status Status { get; set; }
-    public int Number { get; set; }
+    public long Number { get; set; }
 
     //constructor
-    protected Runner(string fullName, string address, int phoneNumber, string emailAddress, RankType rankType, Status status,
-        int number) : base(fullName, address, phoneNumber, emailAddress)
+    protected Runner(string fullName, string address, long phoneNumber, string emailAddress, RankType rankType, Status status,
+        long number) : base(fullName, address, phoneNumber, emailAddress)
     {
         RankType = rankType;
         Status = status;
